26/12/2024

+++ IntelliJ Idea/namcat_e_commerce/src/main/java/vn/com/namcat_e_commerce/controller/GioHangController.java
@@ -59,7 +59,7 @@ public class GioHangController {
         
         String nguoiDung = "";
         int soLuong = 0;
-        long tongTienGH = 0;
+        int tongTienGH = 0;
         
         if(session.getAttribute("user_Online") != null)
         {
